move pool candidate scopes to builder
