+++ api/app/GraphQL/Queries/CountPoolCandidatesByPool.php
@@ -33,10 +33,9 @@ final class CountPoolCandidatesByPool
         });
 
         // available candidates scope (scope CANDIDATE_STATUS_QUALIFIED_AVAILABLE or CANDIDATE_STATUS_PLACED_CASUAL, or PLACED_TENTATIVE)
-        PoolCandidate::scopeAvailable($queryBuilder);
-
+        $queryBuilder->whereAvailable()
         // Only display IT & OTHER publishing group candidates
-        PoolCandidate::scopeInTalentSearchablePublishingGroup($queryBuilder);
+            ->whereInTalentSearchablePublishingGroup();
 
         $queryBuilder->whereHas('user', function ($userQuery) use ($filters) {
             // user filters go here
